requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-GraphQL==2.0.1
graphql-core==2.3.2
graphql-server-core==1.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
promise==2.3
pymongo==4.6.0
python-dotenv==1.0.0
Rx==1.6.3
six==1.16.0
Werkzeug==3.0.1